requirements.txt
Flask==2.1.3
gunicorn==20.1.0
WTForms==3.0.1
Flask-WTF==1.0.1
python-magic==0.4.27
google-cloud-storage==2.4.0